fix(runtime): 64-bit length+offset in read_file_bytes/read_bytes_at for >=4GB files
read_file_bytes(path) stored the byte-buffer length in a 32-bit field, so a file
>= 2^32 bytes (4.294 GB) wrapped modulo 2^32. Repro: anima's 1B ByteGPT mount
binary h1167_1b.bin is 4,325,902,356 bytes; len(rb) returned 30,935,060 and the
buffer was unusable (header u32s read 0 -> d/n_head = 0/0 -> "division by zero"),
blocking 1B+ model mounts. Files < 4.29 GB worked.

Make the whole chain 64-bit (root cause, not a size cap):
  - HexaArr.len/.cap (+ HexaArrI64/F64, preserving the (ptr,len,cap) layout-match
    invariant) and runtime.h's HexaArr typedef -> int64_t (the codegen-emitted
    user.c includes that header; both struct defs must match the runtime ABI).
  - hexa_len() -> int64_t (codegen lowers len(x) -> hexa_int(hexa_len(x)), so
    every caller is safe; the few `int __iter_len` locals narrow as before).
  - rt_read_file_bytes / rt_read_bytes_at: drop the (int) length/offset casts,
    store/iterate with int64_t. Seek/read via the SHIMMED fseek/ftell/fread
    (hxlcl_*, `long` = 64-bit on LP64 over hxlcl_lseek(off_t)) -- the unshimmed
    fseeko/_fseeki64 crash the hxlcl_fopen handle (flockfile SIGSEGV), which is
    why read_bytes_at was already broken on any offset.
  - bytes_to_f32_le/_f64_le/_v: int len -> int64_t len (offset into a >=2 GB
    buffer). OOB-message %d -> %lld for the widened field.

SSOT edits in the emitters (runtime_core_emit.hexa, tensor_kernels_emit.hexa) +
the tracked header runtime.h; runtime_core.c/tensor_kernels.c are gitignored
generated artifacts (regenerated byte-parity from the emitters).

Verify (macOS arm64, LP64):
  - C harness on the real 4.3 GB bin: OLD (int)len = 30935060 (reproduces bug),
    NEW int64 = 4325902356; header u32 = 256,1792,28,16,512.
  - in-engine read_bytes_at on the real bin: header u32 = 256/1792/28/16/512;
    ranged reads at offset 2.2e9 (past 2^31) and 4.3258e9 (past 2^32) succeed.
  - read_file_bytes 100 MB: correct len + high-offset index; binary round-trip
    test PASS; runtime recompiles clean (0 errors).

+++ self/runtime.h
@@ -71,7 +71,7 @@ typedef struct HexaMapTable {
     int          from_arena;
 } HexaMapTable;
 
-typedef struct HexaArr { HexaVal* items; int len; int cap; }            HexaArr;
+typedef struct HexaArr { HexaVal* items; int64_t len; int64_t cap; }    HexaArr; /* 64-bit len/cap: a >=4GB read_file_bytes buffer must not wrap (MUST match runtime_core.c) */
 typedef struct HexaMap { HexaMapTable* tbl; int len; }                  HexaMap;
 typedef struct HexaFn  { void* fn_ptr; int arity; }                     HexaFn;
 typedef struct HexaClo { void* fn_ptr; int arity; HexaVal* env_box; }   HexaClo;
@@ -223,7 +223,7 @@ void    hexa_set_args(int argc, char** argv);         /* runtime.c:5088 */
 HexaVal hexa_args(void);                              /* runtime.c:5103 */
 
 /* arrays */
-int     hexa_len(HexaVal v);                          /* runtime.c:2042 */
+int64_t hexa_len(HexaVal v);                          /* runtime.c:2042 — int64: ≥4GB read_file_bytes len must not truncate */
 HexaVal hexa_array_new(void);                         /* runtime.c:1693 */
 HexaVal hexa_array_push(HexaVal arr, HexaVal item);   /* runtime.c:1809 */
 HexaVal hexa_float(double f);                         /* runtime.c:1232 */
